requirements.txt
pytest>=8.0.0
playwright>=1.41.0
requests>=2.31.0
filelock>=3.13.1
pytest-playwright>=0.4.4
pymongo>=4.6.1
pytest-xdist>=3.5.0
python-dotenv>=1.0.0
